Add TimeSinceUpdate component
