+++ app/src/main/res/drawable/ic_update.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="960"
+    android:viewportHeight="960">
+  <path
+      android:pathData="M480,840q-75,0 -140.5,-28.5t-114,-77q-48.5,-48.5 -77,-114T120,480q0,-75 28.5,-140.5t77,-114q48.5,-48.5 114,-77T480,120q82,0 155.5,35T760,254v-94h80v240L600,400v-80h110q-41,-56 -101,-88t-129,-32q-117,0 -198.5,81.5T200,480q0,117 81.5,198.5T480,760q105,0 183.5,-68T756,520h82q-15,137 -117.5,228.5T480,840ZM592,648L440,496v-216h80v184l128,128 -56,56Z"
+      android:fillColor="#000000"/>
+</vector>
